requirements.txt
google-generativeai
google-cloud-aiplatform
flask
flask-cors
python-dotenv
pymupdf
pydantic
werkzeug